Move lib/auditd to session/auditd
