+++ session/auditd/auditd.go
