+++ session/auditd/common.go
